refactor(core): load SVG exporter dynamically
Move SVG contexts and mark renderers into a lazy feature module.

Keep export-only code out of the ESM entry and update the roadmap.
