+++ packages/core/src/marks/rect.js
@@ -15,16 +15,6 @@ import {
 import { getCachedOrCall } from "../utils/propertyCacher.js";
 import { isDiscrete } from "vega-scale";
 import { cssColorToArray } from "../gl/colorUtils.js";
-import { createSvgElement } from "../view/renderingContext/svgViewRenderingContext.js";
-import {
-    createSvgAttributeEncoder,
-    encodeNumber,
-    encodePosition,
-    formatSvgNumber,
-    projectX,
-    projectY,
-    toSvgString,
-} from "./svgMarkUtils.js";
 
 const hatchPatterns = [
     "none",
@@ -288,83 +278,6 @@ export default class RectMark extends Mark {
                 offset
             );
         }, options);
-    }
-
-    /**
-     * @param {import("../view/renderingContext/svgViewRenderingContext.js").default} context
-     * @param {import("../view/renderingContext/svgViewRenderingContext.js").SvgMarkRenderingOptions} options
-     */
-    renderSvg(context, options) {
-        if (
-            this.#isRoundedCorners() ||
-            (this.properties.hatch ?? "none") != "none" ||
-            this.properties.shadowOpacity
-        ) {
-            throw new Error(
-                "SVG export does not support rounded, hatched, or shadowed rectangles yet."
-            );
-        }
-
-        const { coords, data, group, viewOpacity } = options;
-        const encoders =
-            /** @type {Record<string, import("../types/encoder.js").Encoder>} */ (
-                this.encoders
-            );
-        const encodeStyles = createSvgAttributeEncoder(group, {
-            fill: { encoder: encoders.fill, transform: toSvgString },
-            "fill-opacity": {
-                encoder: encoders.fillOpacity,
-                transform: (value) => +value * viewOpacity,
-            },
-            stroke: { encoder: encoders.stroke, transform: toSvgString },
-            "stroke-opacity": {
-                encoder: encoders.strokeOpacity,
-                transform: (value) => +value * viewOpacity,
-            },
-            "stroke-width": {
-                encoder: encoders.strokeWidth,
-                transform: (value) => formatSvgNumber(+value),
-            },
-        });
-
-        for (const datum of data) {
-            const xOffset = encodeNumber(encoders.xOffset, datum);
-            const yOffset = encodeNumber(encoders.yOffset, datum);
-            const x2Offset = encoders.x2Offset
-                ? encodeNumber(encoders.x2Offset, datum)
-                : xOffset;
-            const y2Offset = encoders.y2Offset
-                ? encodeNumber(encoders.y2Offset, datum)
-                : yOffset;
-            const x1 = projectX(
-                coords,
-                encodePosition(encoders.x, datum),
-                xOffset
-            );
-            const x2 = projectX(
-                coords,
-                encodePosition(encoders.x2, datum),
-                x2Offset
-            );
-            const y1 = projectY(
-                coords,
-                encodePosition(encoders.y, datum),
-                yOffset
-            );
-            const y2 = projectY(
-                coords,
-                encodePosition(encoders.y2, datum),
-                y2Offset
-            );
-            const rect = createSvgElement("rect", {
-                x: formatSvgNumber(Math.min(x1, x2)),
-                y: formatSvgNumber(Math.min(y1, y2)),
-                width: formatSvgNumber(Math.abs(x2 - x1)),
-                height: formatSvgNumber(Math.abs(y2 - y1)),
-                ...encodeStyles(datum),
-            });
-            group.appendChild(rect);
-        }
     }
 
     /**
