+++ packages/core/src/marks/text.js
@@ -15,16 +15,6 @@ import Mark from "./mark.js";
 import { fixCoveragePositional, fixHalfOpenRangedText } from "./markUtils.js";
 import { primaryPositionalChannels } from "../encoder/encoder.js";
 import { requestFont } from "../fonts/textMetrics.js";
-import { createSvgElement } from "../view/renderingContext/svgViewRenderingContext.js";
-import {
-    createSvgAttributeEncoder,
-    encodeNumber,
-    encodePosition,
-    formatSvgNumber,
-    projectX,
-    projectY,
-    toSvgString,
-} from "./svgMarkUtils.js";
 
 /** For GLSL uniforms */
 const alignments = {
@@ -266,126 +256,4 @@ export default class TextMark extends Mark {
             options
         );
     }
-
-    /**
-     * @param {import("../view/renderingContext/svgViewRenderingContext.js").default} context
-     * @param {import("../view/renderingContext/svgViewRenderingContext.js").SvgMarkRenderingOptions} options
-     */
-    renderSvg(context, options) {
-        if (this.properties.logoLetters || this.properties.fitToBand) {
-            throw new Error(
-                "SVG export does not support fitted or logo-letter text yet."
-            );
-        }
-
-        const { coords, data, group, viewOpacity } = options;
-        const encoders =
-            /** @type {Record<string, import("../types/encoder.js").Encoder>} */ (
-                this.encoders
-            );
-        const channelDef = this.encoding.text;
-        const numberFormat =
-            "format" in channelDef
-                ? format(channelDef.format)
-                : (/** @type {any} */ d) => d;
-        const encodeStyles = createSvgAttributeEncoder(group, {
-            fill: { encoder: encoders.color, transform: toSvgString },
-            "fill-opacity": {
-                encoder: encoders.opacity,
-                transform: (value) => +value * viewOpacity,
-            },
-            "font-size": {
-                encoder: encoders.size,
-                transform: (value) => formatSvgNumber(+value),
-            },
-        });
-        group.setAttribute("font-family", "sans-serif");
-        group.setAttribute("font-style", this.properties.fontStyle ?? "normal");
-        group.setAttribute(
-            "font-weight",
-            "" + normalizeFontWeight(this.properties.fontWeight ?? "normal")
-        );
-        group.setAttribute("text-anchor", textAnchors[this.properties.align]);
-        group.setAttribute(
-            "dominant-baseline",
-            dominantBaselines[this.properties.baseline]
-        );
-
-        for (const datum of data) {
-            const value = numberFormat(encoders.text(datum));
-            const stringValue = isString(value)
-                ? value
-                : value === null
-                  ? ""
-                  : "" + value;
-            if (!stringValue) {
-                continue;
-            }
-
-            const x = projectX(
-                coords,
-                encodePosition(encoders.x, datum),
-                encodeNumber(encoders.xOffset, datum)
-            );
-            const y = projectY(
-                coords,
-                encodePosition(encoders.y, datum),
-                encodeNumber(encoders.yOffset, datum)
-            );
-            const size = encodeNumber(encoders.size, datum);
-            const angle = encodeNumber(encoders.angle, datum);
-            const svgX = formatSvgNumber(x);
-            const svgY = formatSvgNumber(y);
-            const text = createSvgElement("text", {
-                x: svgX,
-                y: svgY,
-                dx: formatSvgNumber(this.properties.dx),
-                dy: formatSvgNumber(this.properties.dy),
-                lengthAdjust: "spacingAndGlyphs",
-                textLength: formatSvgNumber(
-                    this.font.metrics.measureWidth(stringValue, size)
-                ),
-                ...encodeStyles(datum),
-            });
-            text.textContent = stringValue;
-            if (angle) {
-                text.setAttribute(
-                    "transform",
-                    `rotate(${formatSvgNumber(angle)} ${svgX} ${svgY})`
-                );
-            }
-            group.appendChild(text);
-        }
-    }
-}
-
-const textAnchors = {
-    left: "start",
-    center: "middle",
-    right: "end",
-};
-
-const dominantBaselines = {
-    top: "text-before-edge",
-    middle: "central",
-    bottom: "text-after-edge",
-    alphabetic: "alphabetic",
-    baseline: "alphabetic",
-};
-
-const fontWeights = {
-    thin: 100,
-    light: 300,
-    regular: 400,
-    normal: 400,
-    medium: 500,
-    bold: 700,
-    black: 900,
-};
-
-/** @param {string | number} weight */
-function normalizeFontWeight(weight) {
-    return typeof weight == "number"
-        ? weight
-        : fontWeights[/** @type {keyof typeof fontWeights} */ (weight)];
 }
