+++ packages/core/src/svg/svgNumber.js
